Extract readSession to ClientSqlCommon
